Change design, prepare for systematic uncertainties
Now the POI to determine is directly the SF, these
way c and l scale factor can float be constraint
to add a source of uncertainty if required.
In addition, MC efficiencies are now a function of
of the normalization. Now normalization factors
which multiply the theory cross sections are used
instead. Builder and PretagTagPdf have been
extended to keep shifted Components, but the
implementation in the evaluation is not done.